implemented led_output module to send RGB565 data to HUB75 display
Implemented an led_output module to send RGB565 data to a HUB75 display,
which uses a series of LED shift registers to display data. RGB565 is
simulated by displaying each bit within the R,G, and B channels for
twice as many scanning periods as the prior bit, to simulate intensity.

Signed-off-by: Daniel DeGrasse <daniel@degrasse.com>